backend/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==3.0.0
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
psycopg2-binary>=2.8
python-decouple==3.8
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
uritemplate==4.1.1
urllib3==1.26.15
